Remove static_labels
The feature may be confusing because the default platform metrics that
are added by the `prometheus_client` library are not changed. This is
intended because those being added are just a side-effect of using the
library.

However, in our use case, we noticed that we don't need those static
labels and rater let outside means (prometheus, kubernetes) add them on
scraping.

No version increment by 1 because the library is alpha anyways and not
published to PyPi